LayoutTests/imported/w3c/web-platform-tests/webxr/hit-test/ar_hittest_subscription_refSpaces.https-expected.txt


PASS Ensures subscription to hit test works with viewer space - straight ahead - plane - webgl
PASS Ensures subscription to hit test works with viewer space - straight ahead - plane - webgl2
PASS Ensures subscription to hit test works with viewer space - straight up - no results - webgl
PASS Ensures subscription to hit test works with viewer space - straight up - no results - webgl2
FAIL Ensures subscription to hit test works with local space - webgl assert_approx_equals: positions must be equal Point comparison failed.
 p1: {x: 0, y: 1, z: -2.5, w: 1}
 p2: {x: 0, y: 0, z: -2.5, w: 1}
 Difference in component y exceeded the given epsilon.
 expected 1 +/- 0.001 but got 0
FAIL Ensures subscription to hit test works with local space - webgl2 assert_approx_equals: positions must be equal Point comparison failed.
 p1: {x: 0, y: 1, z: -2.5, w: 1}
 p2: {x: 0, y: 0, z: -2.5, w: 1}
 Difference in component y exceeded the given epsilon.
 expected 1 +/- 0.001 but got 0
FAIL Ensures subscription to hit test works with local-floor space - webgl assert_approx_equals: positions must be equal Point comparison failed.
 p1: {x: 0, y: 1, z: -2.5, w: 1}
 p2: {x: 0, y: 0.25, z: -2.5, w: 1}
 Difference in component y exceeded the given epsilon.
 expected 1 +/- 0.001 but got 0.25
FAIL Ensures subscription to hit test works with local-floor space - webgl2 assert_approx_equals: positions must be equal Point comparison failed.
 p1: {x: 0, y: 1, z: -2.5, w: 1}
 p2: {x: 0, y: 0.25, z: -2.5, w: 1}
 Difference in component y exceeded the given epsilon.
 expected 1 +/- 0.001 but got 0.25
